pyproject.toml
[project]
name = "chamaleon"
version = "0.1"
description = "A Python package to find sentences matching a given sentiment."
authors = [{ name = "Giskard AI", email = "hello@giskard.ai" }]
dependencies = [
    "transformers>=4.35.2",
    "torch>=2.1.1",
    "wandb>=0.16.1",
    "pandas>=2.1.4",
    "datasets>=2.16.0",
    "trl>=0.7.7",
    "python-Levenshtein>=0.23.0",
]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = { text = "" }

[tool.pdm.dev-dependencies]
dev = ["black>=23.11.0", "ruff>=0.1.6"]
test = ["pytest>=7.4.3"]

[tool.pdm.scripts]
test = "pytest tests"

